Fix #39: Look for exact repository name before using wildcards.
This modifies SourceRepo::load_by_name() to look for a repositoyr with the
exact name given first, and then fallback onto the old method of looking for
a repo that simply contains the given name.  This both addresses the problem
and maintains compatibility with the existing behavior and/or expectations.