Implemented real commit sequence. (Gitweb)
The original commit sequence in SourceGitweb.php was incomplete. Added
code to pull the information out of the data= argument and pull in those
changes. Currently this is the latest commit from gitweb. But it looks
for the commit ID that is passed in then attempts to back fill commits
that are parent of this. It also only attempts to read changesets for
the branch that SourceIntegration knows about.