New SourceHgWeb plugin for Mercurial integration
This plugin is based on the Gitweb plugin as HgWeb/Mercurial are quite
similar to Gitweb/Git. The only missing feature at the moment is the
ability for SourceHgWeb to determine which files were
added/removed/modified with each changeset. The reason is that HgWeb
doesn't give us easy access to this information in the same way Gitweb
does. Hence we will need to manually parse the changeset diff (--- and
+++ lines), checking for /dev/null as a filename in the event of new
files or deleted files.