Fix #51: Improved commit post-processing events
Rather than having separated post-processing events for checkin versus
import, the post-processing has been moved to the API function for
processing commit data.

The first post-process event "commits" allows plugins to get a list of
all changesets as they are checked in or imported.  The second event
"fixed" allows plugins to see a list of bugs that have been resolved by
changesets, along with the appropriate changeset, as described by
feature request #51.

This commit also removes the postcommit/postimport actions from the
MantisSourcePlugin interface, as they were a deemed uneccesary with the
new implementation, and generally should not be needed for VCS support.