Fix #97: Refacter core SVN into shared codebase
Created base SVN plugin with the shared SVN code from WebSVN and SFSVN
plugins, and then modified the previous plugins to inherit from the new
base plugin.  This will end the previous need to constantly replicate
changes between the two SVN plugins.

Currently, the only way to modify shared SVN configuration is to have
the base SVN plugin installed, but this is a task for a different
commit.