Fix incorrect SVN log parsing regexes
When importing a repository with non-standard paths, the parsing regexes
were incorrectly looking for filenames without a leading slash, and also
not looking for filenames with a slash after the branch/tag path.

Also, when importing standard repositories where the trunk/branches/tags
folders were not named in complete lower case, the regexes and logic for
finding those paths has been updated to be case-insensitive, and to
prefer using the same case as found in the repository for trunk.