Fix #144: Proper changeset processing order
When commiting changesets that resolve an issue, the changeset data
should be saved to the database before beginning to modify the issue
data.  This is to fix a problem of generating an email with a change
history that did not show the attached changesets, which could be
confusing in certain cases.