WebSVN : Add option to handle "Multiview" mode
WebSVN can be configured to use Apache Multiview link format [1]. This
commit allows users to specify if this option is enabled or not in the
settings page of their WebSVN repositories.

Note that theses changes only make the option visible in the WebSVN
repos config, but does not alter the plugin's behavior. The next step
is to check the value of this option so the plugin can format WebSVN
links as appropriate.

[1] http://httpd.apache.org/docs/current/content-negotiation.html#multiviews

Fixes #46, #47

Original code was modified to use check_checked() function and fix
whitespace, commmit message reworded.

Signed-off-by: Damien Regad <damien.regad@merckgroup.com>