Partial revert "Replace uses of config_get() by calls to plugin_config_get()"
This partially reverts commit 9222599ab5f04615763262a0b8784d40cf7a2590.

It is required that we use config_get() referencing the 'Source' plugin
in SourceIntegration.php. Because SourceIntegration is a child plugin of
Source, plugin_config_get() does not refer to the correct config string
when retrieving the setting, causing 'Configuration option not found'
errors to occur.