Use svn log with xml output to resolve encoding issues (as reported in issue #62)
As reported in issue #62 there are troubles with svn log and character encoding. It is AFAIK impossible to force an encoding in the svn command line tool for svn log. However, svn log with the --xml option will generate utf-8 encoded output (and is more save and easier to parse IMHO).