Gitlab: handle payload without JSON data
We have a Mantis installation managing projects under SVN, Git with
Gitweb, and now we are installing a Gitlab server.

Problem is the old Git server is using a hook to send POST data not in
JSON format, and this screws the Gitlab plugin, that bypass all other
plugins and try to retrieve changeset from the first repository matching
null repoid (which is the first repo in mantis database, a SVN repo)

Resolving issue by adding a check for null data after the check for empy
payload.

Fixes #121

Original commit message reformatted and reworded

Signed-off-by: Damien Regad <dregad@mantisbt.org>