Gitweb: error handling when commit URL returns no data
Prior to this, the code did a strict-type check of url_get()'s return
value === false, but the function actually returns NULL in case of
failure.

We now check for a non-empty string, which should properly catch any
error encountered by url_get().

In addition, the error message was improved to provide more useful
information to the user.

Fixes #150, #151

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Change from original contribution:
- fixed whitespace
- reworded error message
- detailed commit message