Changeset linking regex should not match empty repo
The regex now checks that the 1st char of the repository name is not
whitespace.

This avoids situations where a string like 'commit: c:repo:sha:' is not
matched properly because it is parsed as repo = ' c', and changeset id =
'repo'.

Also improve the PHPDoc for CHANGESET_LINKING_REGEX.

Fixes #161