Fix changes to author/committer not saved
Prior to this, when editing a changeset, the only change that was
actually applied was when resetting author/committer to the original
value from source control (i.e. selecting '--' from the list). Other
values were not saved.

This was due to use of user_ensure_exists(), which always returns void.

Fixes #163