Only create links for existing bugs
Prior to this, when processing a non-existing bug id referenced in the
commit message, the plugin incorrectly created a link to the issue in
the Source_bug table, and inserted an entry in the bug_history table.
This caused MantisBT timeline to crash as it processed an invalid
history entry.

This commit fixes the incorrect behavior by checking for referenced
bugs' existence in Source_Process_Changesets() function prior to
processing them, so only valid bug links and history records are
created.

Fixes #157