Redirect back to list page after linking issues
Use changeset's revision id as anchor in the changesets list, and use
that as a new 'redirect' parameter for the attach page so it returns
straight to the changeset list's row that called it.

If the new param is not specified, the behavior is as it was before,
i.e. redirect to view page.