Show status color box next to issue id in view page
The view page was not consistently formatting resolved issue IDs; a
regular link was displayed in Attached Issues list, whereas if the same
resolved/closed issue was referenced in the Changeset messages it would
be printed with strikethrough ('resolved' CSS class).

An attempt to fix this by using the same string_get_bug_view_link()
function for the Affected Issues list was not looking good due to the
tooltip with issue description (the same text is already displayed next
to the issue id).

Finally, displaying a box with the issue's status color (similar to the
display in My View and View Issues pages) was implemented.

Fixes #234