Fix change of repo name after full import
If a repository is imported successfully but import_full() does not
actually return any changesets, the temp repo created for the import
remains with its `Import <timestamp>` name instead of being renamed to
the original repo's name, and the latter is deleted.

Setting the temp repo's name right after the initial save operation
prevents this unwanted behavior.

Fixes #245