SVN: pipe handling reordering, read STDOUT first
Due to limitations on the buffering of pipes and therefore possible
resulting deadlocks, the order in which the pipes are read is crucial.

This reordering of the pipes is just a partial workaround. Processing
STDOUT before STDERR will NOT(!) prevent a deadlock if the output to
STDERR exceeds the pipe's max buffered size.

Regardless, as discussed in #261, it is better to handle an unlikely
error scenario, than to frequently get stuck when processing good
commits.

Signed-off-by: Damien Regad <dregad@mantisbt.org>

- Original commit message and comment block reworded
- Updated changelog
- Bump SourceSVN plugin version

Fixes #333