Move the new config to MantisSourceGitBasePlugin
Since it only applies to git-based plugins, the new config should only
be displayed if one of these is installed.

The config has been renamed to 'git_default_primary_branch', and the
name is defined in a class constant to ensure consistency.

A protected method to retrieve the config's value has been added to the
base class, to avoid multiple config_get() calls in each sub-class.

The original language string has been modified; it is no longer useful
to indicate that the config only applies to git-based plugins, because
this is now obvious as it is listed in a git-specific section.