Show "Attach issues" form only if user has access
When a Changeset is not linked to any issue, the List page displays a
text field and an Attach button as a shortcut to link issues without
having to go to the View Changeset page.

The input form was also shown when user has no privileges to update
Changesets.

Fixes #354