Fix missing language strings warnings
On MantisBT 2.25.0, Manage Repository page displays warnings at line 62:
APPLICATION WARNING 300: String "on" not found.
APPLICATION WARNING 300: String "off" not found.

As it turns out, the convert_to_key_value() function introduced in
v2.0.0-beta.2 [[1]] is buggy, and the display value it generates is
never used.

Function was refactored to remove references to the `on` and `off`
language strings and use trans_bool() instead, and to return a properly
escaped string ready for display.

This allows simplification of the code that prints these values.

Fixes #364

[1]: 79497ddbcd5aff9398e05ecc189835a7c22cfa6b