Send a valid security token to print_form_button()
Follow-up on commit 787eaa84ce7a68ead27ceafe2491f27fb6dad707.

Trying to detach an issue from a Changeset triggers
APPLICATION ERROR 2800 - Invalid form security token.

The security token passed on to print_form_button() was not correct -
the function expects an actual token, as generated from function
form_security_token(), and we passed it the token name instead.

A single token is generated outside of the foreach loop that displays
the Issues, and used in the print_form_button() call.