Bitbucket fix changeset retrieval for branches with '/'
This is a workaround for a bug in Bitbucket API [1][2] which returns a
"Resource not found" error when the branch has a '/' in its name.

If the initial commit is a branch name containing a slash, we now look
up the target commit SHA using the refs/branches API endpoint before
entering the main processing loop.

Fixes #376

[1]: https://developer.atlassian.com/cloud/bitbucket/rest/api-group-commits/#api-repositories-workspace-repo-slug-commit-commit-get
[2]: https://jira.atlassian.com/browse/BCLOUD-9969