Only display the Diff/File buttons when needed
In some cases, it does not make any sense to display the Diff or File
buttons (e.g. when a file has been deleted), and the generated target
URL may be invalid (HTTP 404).

The Changeset View page has been adapted to not show the button when the
URL returned by the url_diff() and url_file() methods is empty. It now
behaves like the List page (see Source_View_Changesets()).

Fixes #401