Allow file rename action for GitHub and Bitbucket
To avoid a schema change introducing a new field for the old filename,
we store both the old and new name in the existing field, separated by
a delimiter ' → ' (defined as constant in SourceFile class).

To ensure we generate valid URLs as target for the Diff and File buttons
we also introduce a new SourceFile::getFilename() method, which is used
in the url_diff() and url_file() methods to split the value and return
the actual filename to use (i.e. the new one).

Fixes #374